Update README.md
We have to predict Rossmann Pharmaceutical  wants to forecast sales in all their stores across several cities six weeks ahead of time. Managers in individual stores rely on their years of experience as well as their personal judgment to forecast sales.
The data team identified factors such as promotions, competition, school and state holidays, seasonality, and locality as necessary for predicting the sales across the various stores.
In this project we will discover how to make predictions and deliver and end-to-end product that can help the finance team.